Simplify the handling of line-endings.
This removes a bunch of special-case addition of newlines and attempts to preserve exactly the number provided by the user when inserting or setting.
It also fixes a bug where inserting a line after a 'def' could produce additional newlines after the 'def' block.

Note that this does change the expected results of the setter tests.